% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{SurvivalLogistics}
\alias{SurvivalLogistics}
\title{Logistical Planning for Survival Studies}
\usage{
SurvivalLogistics(
  n_final,
  loss_rate,
  eligibility_rate,
  subjects_per_month,
  follow_up_months,
  annual_event_rate,
  language = "en"
)
}
\arguments{
\item{n_final}{Required sample size}

\item{loss_rate}{Proportion of loss to follow-up (0-1)}

\item{eligibility_rate}{Proportion of eligible people (0-1)}

\item{subjects_per_month}{People that can be recruited per month}

\item{follow_up_months}{Maximum follow-up time in months}

\item{annual_event_rate}{Annual event rate (to estimate when needed events will be reached)}

\item{language}{Language for messages: 'en' (English) or 'es' (Spanish). Default is 'en'}
}
\value{
List with logistical requirements
}
\description{
Calculates logistical requirements for a survival study, considering
recruitment, follow-up, and losses.
}
